Refactor: Restructure files and update imports
