+++ src/app/layout.js
@@ -1,6 +1,7 @@
 import '../styles/globals.scss'
 import '../styles/normalize.css'
-import { ThemeProvider } from '@/components/ThemeContext'
+import { ThemeProvider } from '@/components/contexts/ThemeContext'
+// import { DataProvider } from '@/components/contexts/DataContext'
 import ThemeToggleButton from '@/components/ToggleThemeButton'
 
 export const metadata = {
