+++ src/styles/hero.module.scss
@@ -10,7 +10,7 @@
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center;
-  height: 100vh;
+  height: 108vh;
 }
 
 .heroTitleContainer {
